Changed names from GitRepoExtension to Git2Pages
The previous code used GitRepoExtension to describe functions and classed. I
changed the names to Git2Pages once we decided on the name for the extension
and will be using that name throughout.

Also fixed Git2PagesSetup to work properly. Previously, the code was incomplete
and did not follow mediawiki's format for hook setup.

Change-Id: I7796d03dc59b608a68f2559812e88ed6977cce6b