Added functionality for user inputs.
The user can now add inputs. It will be in the following format:

{{#snippet:repository=MYREPO|file=MYFILE|branch=MYBRANCH|start=START|end=END}}

Position does not matter and currently every input is optional. The code will
recieve all values in $parser and will separately parse through the elements in
function extractOptions.

extractOptions will make a dictionary of the values and will pass it back to
PullContentFromRepo.

Change-Id: Ic443df0dcf6d86157c87c1ae06e100a62b2c6f24