Added git cloning functionality.
Extension will now clone in a unique directory. The folder will be
the result of md5($url). If folder is already present, then the
program will not attempt to reclone from the same url.

Set a higher limit on wfShellExec('git clone...') to allow for larger
git repositories.

Also updated the author of the extension.

Change-Id: Id858ce891acac14afdb29c5956ca31f1bb571e16