Can read and return contents of file.
File can now be read and its contents will be returned depending on user's
startline and endline inputs. If no inputs are found, the entire file contents
will be returned.

Program will also return an error message to the page if filename is not valid
or unreadable.

Changed debug messages to be consistent with the class it's in.

Change-Id: I3fca617ec9375c29773c91a728ec72d530ed078a